backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
websockets==12.0
pydantic==2.9.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
numpy==1.26.0
solders==0.21.0
base58==2.1.1
orjson==3.10.0
